Ajustes

+++ AppMvcCompleta/src/App/Controllers/FornecedoresController.cs
@@ -9,22 +9,27 @@ using System.Threading.Tasks;
 
 namespace App.Controllers
 {
+    [Route("fornecedores/")]
     public class FornecedoresController : Controller
     {
         private readonly IFornecedorRepository _fornecedorRepository;
+        private readonly IEnderecoRepository _enderecoRepository;
         private readonly IMapper _mapper;
 
-        public FornecedoresController(IFornecedorRepository fornecedorRepository, IMapper mapper)
+        public FornecedoresController(IFornecedorRepository fornecedorRepository, IMapper mapper, IEnderecoRepository enderecoRepository)
         {
             _fornecedorRepository = fornecedorRepository;
             _mapper = mapper;
+            _enderecoRepository = enderecoRepository;
         }
 
+        [Route("lista-fornecedores")]
         public async Task<IActionResult> Index()
         {
             return View(_mapper.Map<IEnumerable<FornecedorViewModel>>(await _fornecedorRepository.ObterTodos()));
         }
 
+        [Route("{id}/detalhes-fornecedor")]
         public async Task<IActionResult> Details(Guid id)
         {
             var fornecedorViewModel = await ObterFornecedorEndereco(id);
@@ -37,12 +42,13 @@ namespace App.Controllers
             return View(fornecedorViewModel);
         }
 
-        // GET: Fornecedores/Create
+        [Route("novo-fornecedor")]
         public IActionResult Create()
         {
             return View();
         }
 
+        [Route("novo-fornecedor")]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(FornecedorViewModel fornecedorViewModel)
@@ -58,7 +64,7 @@ namespace App.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        // GET: Fornecedores/Edit/5
+        [Route("{id}/atualizar-fornecedor")]
         public async Task<IActionResult> Edit(Guid id)
         {
             var fornecedorViewModel = await ObterFornecedorProdutosEndereco(id);
@@ -70,6 +76,7 @@ namespace App.Controllers
             return View(fornecedorViewModel);
         }
 
+        [Route("{id}/atualizar-fornecedor")]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(Guid id, FornecedorViewModel fornecedorViewModel)
@@ -86,17 +93,16 @@ namespace App.Controllers
             return RedirectToAction(nameof(Index));
         }
 
+        [Route("{id}/excluir-fornecedor")]
         public async Task<IActionResult> Delete(Guid id)
         {
             var fornecedorViewModel = await ObterFornecedorEndereco(id);
-            if (fornecedorViewModel == null)
-            {
-                return NotFound();
-            }
+            if (fornecedorViewModel == null) return NotFound();
 
-            return View(fornecedorViewModel);
+            return PartialView("_ConfirmaDelete", fornecedorViewModel);
         }
 
+        [Route("{id}/excluir-fornecedor")]
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(Guid id)
@@ -107,8 +113,54 @@ namespace App.Controllers
                 return NotFound();
             }
 
+            var produtos = ObterFornecedorProdutosEndereco(id).Result;
+            if (produtos.Produtos != null)
+            {
+                ViewBag["erroDelete"] = "Nao foi possivel excluir o produto.";
+                return RedirectToAction(nameof(Index));
+            }
+
+            if (fornecedorViewModel.Endereco != null)
+                await _enderecoRepository.Remover(fornecedorViewModel.Endereco.Id);
+
             await _fornecedorRepository.Remover(id);
             return RedirectToAction(nameof(Index));
+        }
+
+        [Route("{id}/obter-endereco-fornecedor")]
+        public async Task<IActionResult> ObterEndereco(Guid id)
+        {
+            var fornecedor = await ObterFornecedorEndereco(id);
+
+            if (fornecedor == null) return NotFound();
+
+            return PartialView("_DetalhesEndereco", fornecedor);
+        }
+
+        [Route("{id}/atualizar-endereco-fornecedor")]
+        public async Task<IActionResult> AtualizarEndereco(Guid id)
+        {
+            var fornecedor = await ObterFornecedorEndereco(id);
+
+            if (fornecedor == null) return NotFound();
+
+            return PartialView("_AtualizaEndereco", new FornecedorViewModel { Endereco = fornecedor.Endereco });
+        }
+
+        [Route("{id}/atualizar-endereco-fornecedor")]
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> AtualizarEndereco(FornecedorViewModel fornecedor)
+        {
+            ModelState.Remove("Nome");
+            ModelState.Remove("Documento");
+            if (!ModelState.IsValid) return PartialView("_AtualizarEndereco", fornecedor);
+
+            await _enderecoRepository.Atualizar(_mapper.Map<Endereco>(fornecedor.Endereco));
+
+            var url = Url.Action("ObterEndereco", "Fornecedores", new { id = fornecedor.Endereco.FornecedorId });
+
+            return Json(new { success = true, url });
         }
 
         private async Task<FornecedorViewModel> ObterFornecedorEndereco(Guid id)
